move dhash in helpers
